pyproject.toml
[project]
name = "stt-whisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-whisper>=20250625",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
]
